DRF/onlineStore/requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
celery
redis